Updated Suduko_Sorting_DFS.py
Added a section to the bottom of the Suduko_Sorting_DFS.py file allowing the user to interact with the program, by entering the total games the user wants the program to go through and solve as well as save the data that it collects. It saves the information to a .CSV file for data review.